# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/",
    "paths": {
        "/projetos/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get All Projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.projeto"
                            }
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Post New Project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Projeto ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Projeto Title",
                        "name": "title",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Projeto Description",
                        "name": "Description",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Projeto equipe",
                        "name": "equipe",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.projeto"
                        }
                    }
                }
            }
        },
        "/projetos/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get Projeto By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Projeto ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.projeto"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.projeto": {
            "type": "object",
            "properties": {
                "Description": {
                    "type": "string"
                },
                "equipe": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}